handle waiting token with null value
Happens when at least two downloads are done: A and B.
A completes successfully, then we need to wait for ~30min.
In order to know the delay, we request a waiting token for B.
The response gives us a delay, but with a waiting token set to null.

Thus, when we receive a null waiting token, we must wait the given delay
and then do a new waiting token request for the same file (B).