Do not re-download files already present on disk
If the file already exists on disk and is the same size of the remote
one, do not re-download it.
If the size is different, the file will be overwritten, as it was
probably aborted.

This addresses #9, meaning that a user can restart multiple times
uptobox-dl with the same links, and it won't process the files that
were already downloaded.

In other words, it can be used as a resume mechanism.